Fixed misaligned memory allocation
Everything was horribly broken after updating Vulkan from 1.2 to 1.3

It turned out to be that the memory allocator was not respecting the requested byte alignment. This seemed to not cause a problem on older Vulkan, but after updating it broke... everything.

This bug was a real pain to track down.